Adding set sketch implementation
Set sketches store a "sketch" of the contents of a set that can
be used to estimate certain properties of the set. In this case,
we use the HyperLogLog sketch of Flajolet, Fusy, Gandouet, and
Meunier to estimate cardinalities of sets, their unions, and
their intersections.

'KADD key member' adds member to key's set. The cardinality of a
key can then be queried with 'KCARD key' and you'll get an
approximation of the cardinality of key that's usually within
2% of the actual cardinality. The KUNIONCARD and KINTERCARD commands
take multiple keys and estimate either the cardinality of their
unions or intersections, respectively. All operations are O(1).