Implementing RDB persistence and AOF rewrite ability in set sketches
Set sketches now take up more space because they at most 2048 members - these
"max hashes" are the members that created the longest run of zeros in a hash
value for every bucket in the HyperLogLog data structure. Storing this extra
information allows us to implement AOF rewrite correctly, since the state of
the HyperLogLog sketch can be recreated with this set of at most 2048 strings.